+++ test/react.test.js
@@ -134,7 +134,7 @@ test("State renders correctly and updates when modified", t => {
     </StorageProvider>);
 
   t.is(container.outerHTML, "<div><p>initial</p><a>Foo</a></div>");
-  t.deepEqual(s.getSnapshot(), { state: { data: "initial", id: "state", nested: {}, params: { data: "initial" } } });
+  t.deepEqual(s.getSnapshot(), { state: { data: "initial", id: "state", nested: {} } });
 
   const link = getByText("Foo");
   t.not(link, undefined);
@@ -144,7 +144,7 @@ test("State renders correctly and updates when modified", t => {
 
   t.is(container.outerHTML, "<div><p>the new one</p><a>Foo</a></div>");
   t.is(link.outerHTML, "<a>Foo</a>");
-  t.deepEqual(s.getSnapshot(), { state: { data: "the new one", id: "state", nested: {}, params: { data: "initial" } } });
+  t.deepEqual(s.getSnapshot(), { state: { data: "the new one", id: "state", nested: {} } });
   t.is(emit.calls.length, 4);
   t.deepEqual(emit.calls[0].arguments, ["stateCreated", ["state"], { data: "initial" }, "initial"]);
   t.deepEqual(emit.calls[1].arguments, ["messageQueued", { tag: "data", data: "the new one" }, ["state", "$"]]);
@@ -206,7 +206,7 @@ test("State is removed when the Provider is unmounted", t => {
     </StorageProvider>);
 
   t.is(container.outerHTML, `<div></div>`);
-  t.deepEqual(s.getSnapshot(), { state: { data: "my initial", id: "state", nested: {}, params: { data: "my initial" } } });
+  t.deepEqual(s.getSnapshot(), { state: { data: "my initial", id: "state", nested: {} } });
 
   rerender(
     <StorageProvider storage={s} />
@@ -230,7 +230,7 @@ test("State is reused at the same level", t => {
     </StorageProvider>);
 
   t.is(container.outerHTML, `<div><p>my initial</p><p>my initial</p></div>`);
-  t.deepEqual(s.getSnapshot(), { state: { data: "my initial", id: "state", nested: {}, params: { data: "my initial" } } });
+  t.deepEqual(s.getSnapshot(), { state: { data: "my initial", id: "state", nested: {} } });
   t.is(emit.calls.length, 1);
   t.deepEqual(emit.calls[0].arguments, ["stateCreated", ["state"], { data: "my initial" }, "my initial"]);
 });
@@ -257,7 +257,7 @@ test("State updates during rendering are respected", t => {
     </StorageProvider>);
 
   t.is(container.outerHTML, `<div><p>updated</p></div>`);
-  t.deepEqual(s.getSnapshot(), { state: { data: "updated", id: "state", nested: {}, params: { data: "my initial" } } });
+  t.deepEqual(s.getSnapshot(), { state: { data: "updated", id: "state", nested: {} } });
   t.is(emit.calls.length, 4);
   t.deepEqual(emit.calls[0].arguments, ["stateCreated", ["state"], { data: "my initial" }, "my initial"]);
   t.deepEqual(emit.calls[1].arguments, ["messageQueued", { tag: "data", data: "updated" }, ["state", "$"]]);
@@ -278,7 +278,7 @@ test("State is removed when the Provider is the last to be unmounted", t => {
   t.is(container.outerHTML,
     `<div></div>`);
   t.deepEqual(s.getSnapshot(),
-    { state: { data: "my initial", id: "state", nested: {}, params: { data: "my initial" } } });
+    { state: { data: "my initial", id: "state", nested: {} } });
 
   rerender(
     <StorageProvider storage={s}>
@@ -287,7 +287,7 @@ test("State is removed when the Provider is the last to be unmounted", t => {
 
   t.is(container.outerHTML, `<div></div>`);
   t.deepEqual(s.getSnapshot(),
-    { state: { data: "my initial", id: "state", nested: {}, params: { data: "my initial" } } });
+    { state: { data: "my initial", id: "state", nested: {} } });
 
   rerender(<StorageProvider storage={s} />);