+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Renamed `ModelDataType` and `ModelInitType` to `TypeofModelData` and
   `TypeofModelInit`.
+### Removed
+- Removed `params` data from `State` objects and snapshot.
 
 ## [0.4.0] - 2019-10-10
 ### Added